Integrated_Scheduling_System-master/appointment_scheduling/requirements-minimal.txt
# Minimal requirements for Django backend - Python 3.13 compatible
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Optional but useful
django-extensions==3.2.3
